feat: Add initial support for cron into CLI
